- Breaking change: Remove `$.fn.multipleSelect.defaults` in favor of possible removal     of jquery dep. (add defaults within MultipleSelect file) - Optimization: Use `str.normalize` where available - Refactoring: Move sprintf, removeDiacritics, and MultipleSelect into separate source files - Refactoring: Implement sprintf as tagged template and use templates more - Linting: ESLint fixes - npm: Ignore test and rollup files - npm: Update devDeps
